requirements.txt
Pillow
opencv-python
sentence-transformers
numpy
matplotlib